refactor : 딜러의 수익 계산 기능을 Dealer -> PlayerProfitStatistics 클래스로 이관
