+++ src/test/java/domain/user/DealerTest.java
@@ -107,24 +107,6 @@ class DealerTest {
         assertThat(isBust).isTrue();
     }
 
-    @DisplayName("플레이어의 수익 통계로부터 딜러의 수익을 계산하는 기능을 테스트한다.")
-    @Test
-    void getDealerProfitFromPlayerProfitStatisticsTest() {
-        // given
-        Dealer dealer = new Dealer();
-        Map<Player, Profit> statistics = new HashMap<>();
-        Player player = new Player(new PlayerName("player"), new BettingMoney(new BigDecimal("1000")));
-        Profit playerProfit = new Profit(new BigDecimal("1000"));
-        statistics.put(player, playerProfit);
-        PlayerProfitStatistics playerProfitStatistics = new PlayerProfitStatistics(statistics);
-
-        // when
-        Profit dealerProfit = dealer.getDealerProfitFromPlayerProfitStatistics(playerProfitStatistics);
-
-        // then
-        assertThat(dealerProfit).isEqualTo(new Profit(new BigDecimal("-1000")));
-    }
-
     @DisplayName("딜러의 상태 결정 기능을 테스트한다.")
     @ParameterizedTest
     @MethodSource("provideCardsAndStatus")
